# More efficient QC.shuffle in transaction generator