Resolves JIRA issue CAD-2062

The transaction generator was "QuickCheck shuffling" potentially large lists - these shuffles have been replaced by more efficient ruffles.