@mrBliss 2020-10-21
@uroboros Great! We've been wanting this in consensus for a while now. Do you have measurements of how much faster this is than the previous version?