@uroboros 2020-10-21
> @uroboros Great! We've been wanting this in consensus for a while now. Do you have measurements of how much faster this is than the previous version?

@mrBliss I was using Tim's genTx benchmark as reference, which looks very promising! (The change was noticeable in the property test times too, but much less pronounced) 

BEFORE -------------------------------

benchmarking gen/genTx/1000
time                 1.537 ms   (1.522 ms .. 1.555 ms)
                     0.999 R²   (0.999 R² .. 0.999 R²)
mean                 1.541 ms   (1.529 ms .. 1.554 ms)
std dev              44.23 μs   (35.12 μs .. 58.08 μs)
variance introduced by outliers: 17% (moderately inflated)

AFTER -------------------------------

benchmarking gen/genTx/1000
time                 837.4 μs   (825.8 μs .. 846.9 μs)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 848.4 μs   (839.9 μs .. 857.3 μs)
std dev              29.36 μs   (24.47 μs .. 35.82 μs)