@uroboros 2020-10-21
would be interesting to see how it plays out in consensus tests