@mrBliss 2020-10-21
So a 2x speed-up. 

At the moment we have disabled the generator in the consensus because it was prohibitively slow, but also because it triggered a failing assertion in the network layer (that the network team isn't going to fix soon). The latter means that we can't turn it back on again :slightly_frowning_face:.

However, we also use this generator (in)directly, when generating random blocks and transactions for our serialisation roundtrip tests, which are rather slow at the moment. If these get 2x faster, that would also be nice.

I'll report back when I have results (this will take a while, as I still have to propagate a bunch of changes).